pyproject.toml
[tool.poetry]
name = "todo_list"
version = "0.1.0"
description = "A todo list using Django Rest Framework."
authors = ["Shizuku Asami <xv.chronos@gmail.com>"]
license = "MIT"
repository = "https://github.com/Shizuku-Asami/todolist"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
django-restframework = "^0.0.1"
django-cors-headers = "^3.11.0"
django-environ = "^0.8.1"
pytest = "^7.1.1"
pytest-django = "^4.5.2"
djangorestframework-simplejwt = "^5.1.0"
psycopg2-binary = "^2.9.3"
uvicorn = "^0.17.6"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
